fix: UPDATE/DELETE/POST company/jobs

+++ backend/company/controllers/handleJobs.js
@@ -55,29 +55,74 @@ const getJobs = async (req, res) => {
     }
 }
 
-const getJobById = async (job_id) => {
-    const job = await prisma.jobs.findUnique({
-        where: { jobId: job_id },
+const getJobById = async (req,res) => {
+    const job_id = Number(req.params.id);
+    const companyEmail = req.companyEmail;
+
+    const company = await prisma.companies.findFirst({
+        where: { email: companyEmail }
     });
-    if (!job) {
-        throw new Error('Job not found');
+    if (!company) {
+        return res.status(404).json({ message: "Company not found" });
     }
-    return job;
+    const companyId = company.companyId;
+    const job = await prisma.jobs.findFirst({
+        where: { jobId: job_id,companyId:companyId },
+        include : { skills: {
+      select: { skillName: true }
+    }}
+    });
+    console.log("Fetched Job:", job);
+    return res.status(200).json({message:"Job fetched Successfully",data : job})
 }
 
-const updateJob = async (job_id,updateData) => {
+const updateJob = async (req,res) => {
+    const job_id = Number(req.params.id);
+    const updateData = req.body;
+    const companyEmail = req.companyEmail;
+    console.log("Update Data:", updateData);
+
+    const company = await prisma.companies.findFirst({
+        where: { email: companyEmail }
+    });
+    if (!company) {
+        return res.status(404).json({ message: "Company not found" });
+    }
+    const companyId = company.companyId;
     const updatedJob = await prisma.jobs.update({
-        where: { jobId: job_id },
-        data: updateData,
+        where: { jobId: job_id,companyId:companyId },
+        data: {
+            jobTitle: updateData.jobTitle,
+            isActive: true,
+            stipend: updateData.stipend,
+            description: updateData.description,
+            skills: {
+                connectOrCreate: updateData.skills.map(skill => ({
+                where: { skillName: skill },
+                create: { skillName: skill }
+                }))
+        }
+    },
+    include: { skills: true }
     });
-    return updatedJob;
+    return res.status(200).json({message:"Successfully updated",data : updatedJob})
 }
 
-const deleteJob = async (job_id) => {
-    await prisma.jobs.delete({
-        where: { jobId: job_id },
+const deleteJob = async (req,res) => {
+    const job_id = Number(req.params.id);
+    const companyEmail = req.companyEmail;
+
+    const company = await prisma.companies.findFirst({
+        where: { email: companyEmail }
     });
-    return;
+    if (!company) {
+        return res.status(404).json({ message: "Company not found" });
+    }
+    const companyId = company.companyId;
+    await prisma.jobs.delete({
+        where: { jobId: job_id ,companyId:companyId},
+    });
+    return res.status(200).json({message:"Job Deleted Successfully"})
 }
